fix: update test imports after API file rename
Update test file imports to use .client suffix after renaming:
- admin.test.ts: ../admin -> ../admin.client
- refresh.property.test.ts: ../../api/auth -> ../../api/auth.client
- batch-cache-population.property.test.ts: ../../api/user-actions -> ../../api/user-actions.client
