+++ apps/web-client/src/core/auth/__tests__/refresh.property.test.ts
@@ -10,7 +10,7 @@ import {
   _resetRefreshState,
 } from '../refresh';
 import { tokenStorage } from '../token-storage';
-import { authApi } from '../../api/auth';
+import { authApi } from '../../api/auth.client';
 import {
   broadcastRefreshStart,
   broadcastRefreshSuccess,
