+++ apps/web-client/src/core/saved-status/__tests__/batch-cache-population.property.test.ts
@@ -4,7 +4,7 @@
 
 import * as fc from 'fast-check';
 import { queryKeys, createBatchHash } from '../../query/keys';
-import type { MediaSaveStatusDto } from '../../api/user-actions';
+import type { MediaSaveStatusDto } from '../../api/user-actions.client';
 
 describe('Batch to individual cache population', () => {
   /**
